feat: session share
