+++ apps/ui-tars/src/renderer/src/components/ChatInput/index.tsx
@@ -17,7 +17,7 @@ import {
   TooltipTrigger,
 } from '@renderer/components/ui/tooltip';
 import { Button } from '@renderer/components/ui/button';
-import { useScreenRecord } from '@renderer/hooks/useScreenRecord';
+// import { useScreenRecord } from '@renderer/hooks/useScreenRecord';
 import { api } from '@renderer/api';
 
 import { Play, Send, Square, Loader2 } from 'lucide-react';
@@ -47,7 +47,7 @@ const ChatInput = () => {
   };
 
   const { run } = useRunAgent();
-  const { startRecording, stopRecording, recordRefs } = useScreenRecord();
+  // const { startRecording, stopRecording, recordRefs } = useScreenRecord();
 
   const { currentSessionId, updateSession, createSession } = useSession();
 
@@ -57,9 +57,9 @@ const ChatInput = () => {
   // console.log('running', 'status', status, running);
 
   const startRun = async () => {
-    startRecording().catch((e) => {
-      console.error('start recording failed:', e);
-    });
+    // startRecording().catch((e) => {
+    //   console.error('start recording failed:', e);
+    // });
     const instructions = getInstantInstructions();
 
     console.log('startRun', instructions, restUserData);
@@ -108,22 +108,24 @@ const ChatInput = () => {
 
   const isCallUser = useMemo(() => status === StatusEnum.CALL_USER, [status]);
 
+  // console.log('status', status);
+
   /**
    * `call_user` for human-in-the-loop
    */
-  useEffect(() => {
-    // if (status === StatusEnum.CALL_USER && savedInstructions) {
-    //   setLocalInstructions(savedInstructions);
-    // }
-    // record screen when running
-    if (status !== StatusEnum.INIT) {
-      stopRecording();
-    }
+  // useEffect(() => {
+  //   // if (status === StatusEnum.CALL_USER && savedInstructions) {
+  //   //   setLocalInstructions(savedInstructions);
+  //   // }
+  //   // record screen when running
+  //   if (status !== StatusEnum.INIT) {
+  //     stopRecording();
+  //   }
 
-    return () => {
-      stopRecording();
-    };
-  }, [status]);
+  //   return () => {
+  //     stopRecording();
+  //   };
+  // }, [status]);
 
   const lastHumanMessage =
     [...(messages || [])]
@@ -211,10 +213,10 @@ const ChatInput = () => {
         </div>
       </div>
 
-      <div style={{ display: 'none' }}>
+      {/* <div style={{ display: 'none' }}>
         <video ref={recordRefs.videoRef} />
         <canvas ref={recordRefs.canvasRef} />
-      </div>
+      </div> */}
     </div>
   );
 };
