+++ apps/ui-tars/src/renderer/src/hooks/useScreenRecord.ts
@@ -195,11 +195,12 @@ export const useScreenRecord = (
   };
 
   const canSaveRecording = !isRecording && recordedChunksRef.current.length > 0;
-  // console.log(
-  //   '[canSaveRecording]',
-  //   isRecording,
-  //   recordedChunksRef.current.length,
-  // );
+  console.log(
+    '[canSaveRecording]',
+    canSaveRecording,
+    isRecording,
+    recordedChunksRef.current.length,
+  );
 
   return {
     isRecording,
